Bug 2067753 - cargo clippy --fix pass in style. r=firefox-style-system-reviewers,dshin
I discarded some of the more dubious changes manually.

Differential Revision: https://phabricator.services.mozilla.com/D322321

+++ style/values/computed/length.rs
@@ -277,7 +277,7 @@ impl ToTyped for CSSPixelLength {
     fn to_typed(&self, dest: &mut ThinVec<TypedValue>) -> Result<(), ()> {
         dest.push(TypedValue::Numeric(NumericValue::Unit(UnitValue {
             numeric_type: NumericType::length(),
-            value: self.0 as f32,
+            value: self.0,
             unit: CssString::from("px"),
         })));
         Ok(())
